Move run ingest health projection inward
